+++ src/rpc/net.cpp
@@ -1077,6 +1077,9 @@ static RPCHelpMan getnodeaddresses() {
             {"count", RPCArg::Type::NUM, /* default */ "1",
              "The maximum number of addresses to return. Specify 0 to return "
              "all known addresses."},
+            {"network", RPCArg::Type::STR, "all networks",
+             "Return only addresses of the specified network. Can be one of: " +
+                 Join(GetNetworkNames(), ", ") + "."},
         },
         RPCResult{RPCResult::Type::ARR,
                   "",
@@ -1101,7 +1104,11 @@ static RPCHelpMan getnodeaddresses() {
                        }},
                   }},
         RPCExamples{HelpExampleCli("getnodeaddresses", "8") +
-                    HelpExampleRpc("getnodeaddresses", "8")},
+                    HelpExampleCli("getnodeaddresses", "4 \"i2p\"") +
+                    HelpExampleCli("-named getnodeaddresses",
+                                   "network=onion count=12") +
+                    HelpExampleRpc("getnodeaddresses", "8") +
+                    HelpExampleRpc("getnodeaddresses", "4, \"i2p\"")},
         [&](const RPCHelpMan &self, const Config &config,
             const JSONRPCRequest &request) -> UniValue {
             NodeContext &node = EnsureNodeContext(request.context);
@@ -1118,9 +1125,19 @@ static RPCHelpMan getnodeaddresses() {
                                    "Address count out of range");
             }
 
+            const std::optional<Network> network{
+                request.params[1].isNull()
+                    ? std::nullopt
+                    : std::optional<Network>{
+                          ParseNetwork(request.params[1].get_str())}};
+            if (network == NET_UNROUTABLE) {
+                throw JSONRPCError(RPC_INVALID_PARAMETER,
+                                   strprintf("Network not recognized: %s",
+                                             request.params[1].get_str()));
+            }
             // returns a shuffled list of CAddress
-            const std::vector<CAddress> vAddr{node.connman->GetAddresses(
-                count, /* max_pct */ 0, /* network */ std::nullopt)};
+            const std::vector<CAddress> vAddr{
+                node.connman->GetAddresses(count, /* max_pct */ 0, network)};
             UniValue ret(UniValue::VARR);
 
             for (const CAddress &addr : vAddr) {